+++ visualization/app/codeCharta/ui/toolBar/toolBar.module.ts
@@ -11,6 +11,7 @@ import { HoveredNodePathPanelComponent } from "./hoveredNodePathPanel/hoveredNod
 import { FilePanelComponent } from "../filePanel/filePanel.component"
 import { ScreenshotButtonComponent } from "../screenshotButton/screenshotButton.component"
 import { CopyToClipboardButtonComponent } from "../copyToClipboardButton/copyToClipboardButton.component"
+import { DownloadButtonComponent } from "./downloadButton/downloadButton.component"
 
 angular
 	.module("app.codeCharta.ui.toolBar", ["app.codeCharta.state", "app.codeCharta.ui.dialog"])
@@ -22,3 +23,4 @@ angular
 	.directive("ccScreenshotButton", downgradeComponent({ component: ScreenshotButtonComponent }))
 	.directive("ccFilePanel", downgradeComponent({ component: FilePanelComponent }))
 	.directive("ccCopyToClipboardButton", downgradeComponent({ component: CopyToClipboardButtonComponent }))
+	.directive("ccDownloadButton", downgradeComponent({ component: DownloadButtonComponent }))